+++ src/app/market/(mobile)/layout.mobile.tsx
@@ -5,7 +5,6 @@ import { PropsWithChildren } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import AppLayoutMobile from '@/layout/AppLayout.mobile';
-import { useSwitchSideBarOnInit } from '@/store/global';
 import { SidebarTabKey } from '@/store/global/initialState';
 
 import Header from './features/Header';
@@ -13,10 +12,8 @@ import Header from './features/Header';
 const DetailModal = dynamic(() => import('./features/AgentDetail'));
 
 const MarketLayout = ({ children }: PropsWithChildren) => {
-  useSwitchSideBarOnInit(SidebarTabKey.Market);
-
   return (
-    <AppLayoutMobile navBar={<Header />} showTabBar>
+    <AppLayoutMobile navBar={<Header />} showTabBar tabBarKey={SidebarTabKey.Market}>
       <Flexbox flex={1} gap={16} style={{ padding: 16 }}>
         {children}
       </Flexbox>