+++ src/store/global/slices/common/action.ts
@@ -12,8 +12,9 @@ import type { GlobalStore } from '@/store/global';
 import type { GlobalServerConfig, GlobalSettings } from '@/types/settings';
 import { merge } from '@/utils/merge';
 import { setNamespace } from '@/utils/storeDebug';
+import { switchLang } from '@/utils/switchLang';
 
-import type { SidebarTabKey } from './initialState';
+import { settingsSelectors } from '../settings/selectors';
 
 const n = setNamespace('common');
 
@@ -23,7 +24,6 @@ const n = setNamespace('common');
 export interface CommonAction {
   refreshUserConfig: () => Promise<void>;
   switchBackToChat: (sessionId?: string) => void;
-  switchSideBar: (key: SidebarTabKey) => void;
   updateAvatar: (avatar: string) => Promise<void>;
   useCheckLatestVersion: () => SWRResponse<string>;
   useFetchServerConfig: () => SWRResponse;
@@ -44,9 +44,7 @@ export const createCommonSlice: StateCreator<
   switchBackToChat: (sessionId) => {
     get().router?.push(SESSION_CHAT_URL(sessionId || INBOX_SESSION_ID, get().isMobile));
   },
-  switchSideBar: (key) => {
-    set({ sidebarKey: key }, false, n('switchSideBar', key));
-  },
+
   updateAvatar: async (avatar) => {
     await userService.updateAvatar(avatar);
     await get().refreshUserConfig();
@@ -87,6 +85,11 @@ export const createCommonSlice: StateCreator<
           if (!data) return;
 
           set({ avatar: data.avatar, settings: data.settings }, false, n('fetchUserConfig', data));
+
+          const { language } = settingsSelectors.currentSettings(get());
+          if (language === 'auto') {
+            switchLang('auto');
+          }
         },
         revalidateOnFocus: false,
       },